requirements.txt
transformers
pre-commit
torch
open-interpreter
gitpython
pytest
pytest-cov
pytest-mock
pytest-watch
isort
yamllint
black
pylint
flake8
python-dateutil
setuptools>=65.5.1
huggingface_hub
protobuf
sentencepiece
